featt(http2): Update service endpoints
Currently, gRPC traffic to Istio-enabled SeldonDeployments does not work
with the new executor.  This is due to fact that pilot does not allow
for a second listener to forward to the same targetPort, causing all
traffic destined for the shared targetPort to use the first protocol
registered to pilot (which is usually http since it is defined first in
the service).  This updates the service to only create a single service
port using the http2 protocol, which should allow both gRPC and http
traffic through the istio-proxy sidecar and through to the executor.

Contributes to SeldonIO/seldon-core#2189

Signed-off-by: Nick Groszewski <groszewn@gmail.com>