Update flask_utils.py
In the original implementation the binData could only receive one binary file (i.e. image). In this way the get_multi_form_data_request() iterates through all the elements in the request with the "binData" key and append it to the req_dict[fileKey] dictionary, so multiple images can be sent to the application.