Remove is_proto check in send_feedback()
By type definition of the request parameter in the input signature, `is_proto` was always incorrectly set to False. This resulted in `handle_raw_custom_metrics` acting as a no-op.

This PR removes the `isinstance` check on the request parameter since it is always true unless the caller is not respecting argument types. If the caller is not respecting argument types, the method would crash anyway upon calling `seldon_metrics.update_reward(request.reward)` thus I suspect there is no need to protect against such a scenario for reverse compatibility.