Use default PID not UUID for worker ID
The use of a UUID means every request is treated as a new timeseries
by Prometheus, because the labels are different every time.
Furthermore, it means metrics, whether gauges or counters,
can never be reset, which is painful for alerting.
The default worker ID function--the process ID for the worker--
will be consistent across requests, which mitigates both the
aforementioned issues.