Remove option pattern for OAuth store config
This was overcomplicated, convoluted, and less efficient for this use case.
In cases where enabling an option involves non-trivial computation and/or access to private data, this pattern makes sense.
In this case, however, it was used to pass on options unmodified, which is more simply represented with a config struct.