bump(deps): grpc to 1.61, protobuf to 3.25.2, kotlinx-coroutines to 1.7.3
Upgrading grpc introduced a breaking change, where dataflow was not
able to connect to the scheduler at runtime due to a name resolution
failure.

The problem appeared due to a naming resolution code rewrite, which
made it modular and reliant on the implementation being loaded at
runtime by the JVM, from the jar's META-INF/services.

Both grpc (core) and grpc-netty-shaded wrote the information about
the naming resolution implementation to the same path within their
packages. When building the dataflow fat jar, only one implementation
ended up being available, the one for naming resolution of URIs for
unix domain sockets.

The solution was to use the ShadowJdk gradle plugin and configure it
to merge files with the same name in META-INF/services coming from
multiple dependencies.

Other version bumps:
  - kubernetes:client-java
  - junit-jupyter-params