fix(dataflow/gradle): change dependency ordering
The order in which dependencies are specified for Gradle matters,
influencing the ordering in the classpath of the application.

Here, the slf4j klogging plugin was introduced after the kafka
streams dependency, leading to errors at runtime, with the kafka-streams
logging facility not able to find a slf4j provider.