fix(dataflow): clear kafka stream state for failed pipelines
Fixes INFRA-755 (internal): Exceptions in pipeline creation/deletion are
unrecoverable because of uncleaned Kafka Streams state

Because we're now continuously retrying the subscription to the scheduler
when the subscription terminates, we need to also clean the kafka streams
state of failed pipelines when the subscription is terminated with an error.

This commit also introduces a new way of handling errors and the changes
in pipeline status. It moves us towards being able to inspect the status of the
status of the underlying kafka streams from outside the Pipeline object.
The end-goal (not archieved here) is to react to pipelines reaching error
states during their operation, rather than just when they are created for the
first time (what's currently implemented).

Instead of propagating exceptions, the idea is to return and handle common
errors in a more similar way to golang.