fix(dataflow) wait for kafka topic creation
Kafka topic creation happens asynchronously. This means that even when
the return value from `createTopics(...)` indicates that the topic has
been created successfuly, the topic can not be immediately subscribed
to.

Instead of verifying the status of the topic from the `createTopics`
return value, here we're repeatedly calling `describeTopics` until
all of the topics for the pipeline can be described successfully.
This indicates that the topic has been fully created _at least_ on
one broker, and can now be subscribed to.