fix(dataflow): handle pipeline errors and clear kafka streams state (#5358)
We need better error handling and reporting across dataflow, together with 
appropriately cleaning the kafka streams state once a pipeline reaches an
error state.

Because we're now continuously retrying the subscription to the scheduler 
when the subscription terminates, not cleaning the kafka streams state of
failed pipelines may lead to the inability to re-create the same pipeline, even
after deletion.

This commit also introduces a new way of handling errors and the changes
in pipeline status. It moves us towards being able to inspect the status of the
underlying kafka streams from outside the Pipeline object. 

The end-goal (not archieved here) is to react to pipelines reaching error states
during their operation, rather than just when they are created for the first time 
(what's currently implemented). 

However, the changes here do get us to a point where the pipeline creation status,
including possible errors/exceptions are reported to the scheduler and can show 
up either in the seldon CLI or in k8s.

Instead of propagating exceptions, the idea is to return and handle common errors
in a more similar way to golang.

**Fixed issues**:

- Fixes #INFRA-755 (internal): Exceptions in pipeline creation/deletion are 
  unrecoverable because of uncleaned Kafka Streams state
- Progress on #INFRA-617 (internal): Investigate high priority unhandled errors
- Progress on #INFRA-648 (internal): Handle dataflow [...] exceptions and present 
  meaningful error messages