fix(gateway): wait for kafka topic creation (#5359)
* fix(gateway) wait for kafka topic creation

Kafka topic creation happens asynchronously. This means that even when the
return value from createTopics(...) indicates that the topic has been created
successfuly, the topic can not be immediately subscribed to.

We retry DescribeTopics until all of the topics for the pipeline can be
described successfully. This indicates that the topic has been fully created at
least on one broker, and can now be subscribed to.

Minor changes:
- timeout config now in gateway/constants.go
- tidying-up of error reporting

Which issue(s) this PR fixes:
Fixes gateway component for #INFRA-663 (internal): Pipeline creation goes into ERROR state