fix(dataflow) be more strict about error states on pipeline creation (#5381)
Only return success = true for a pipeline create operation if the pipeline has started and the corresponding kafka streams are running

A concurrency issue remains because if running the code on multiple coroutines
these may interleave at suspension points. This is not an issue atm, as we are
executing on one coroutine. If/when we introduce more parallelism, this will
need to be dealt with.

**Fixed issues**:
- #INFRA-830 (internal): pipeline might be marked ready before is actually running