fix(modelgateway): kafka topics with correct number of partitions
The number of partitions with which model-specific kafka topics
is created needs to be the same as the number of partitions for
other topics (pipelines/experiments/etc). This is a Kafka Streams
constraint in order for joins to work correctly.

Until now, an incorrect environment variable was being passed to
modelgateway, which meant that it was always creating topics with
1 partition. This in turn led to dataflow processing getting stuck
whenever other topics were created with multiple partitions

**Fixed issues**:
- INFRA-822: Pipelines get stuck on triggers/joins