fix(dataflow) CVEs related to com.microsoft.azure:adal4j
com.microsof.azure:adal4j is a deprecated library, and an _optional_ dependency
of io.kubernetes:client-java.

However, when building fat-jars, gradle and the "shadow" plugin fetches all
transitive depdendencies, including optional ones (i.e ones which are not used).

In this case, we don't use any of the code in adal4j via io.kubernetes.