build(pb, grpc): update Makefile responsible for code generation
Allow protoc and the language-specific plugins to be installed at specific
versions locally and then use those versions to generate code.

This means we no longer depend on arbitrary versions installed on the
system doing the code generation, making this reproducible.

The installs happen in the apis/tools directory, if the required tools
are not found. Python packages are installed in a virtual environment
under apis/tools/python

In this first phase, we retain the exact same tool versions that were
used to generate the existing protobuf/grpc code. We also apply the
licensing headers to those generated files automatically.

Because of a previous mismatch in how those licensing headers were
generated, all generated files end up with a minor whitespace difference.