build(grpc): update Makefile responsible for code generation (#5749)
Allow protoc and the language-specific plugins to be installed at specific versions locally and then use those versions to generate code.

This means we no longer depend on arbitrary versions installed on the system doing the code generation, making this reproducible.

The installs happen in the `apis/tools` directory, if the required tools are not found. Python packages are installed in a virtual environment under `apis/tools/python`

In this first phase, we retain the exact same tool versions that were used to generate the existing protobuf/grpc code. We also apply the licensing headers to those generated files automatically.

Because of a previous mismatch in how those licensing headers were generated, most generated files end up with a minor whitespace difference.

[IMPORTANT]: The kotlin/java codegen for `apis/mlops/v2_dataplane/v2_dataplane.proto` had not been previously generated using the makefile (it seems it was generated as a one-off). This means the version of the codegen tools used was already mismatched compared to the rest. Consequently, this PR adds this code generation to the Makefile, and re-generates the code files in order to bring everything to a consistent version.

**Issue(s) related to this PR**:

- INFRA-1076 - facilitate protobuf/gRPC codegen version updates