fix(agent): cleanup rclone directory on errors
Previously, any error occurring after the downloading of the model artifact but
before the model was properly registered in the agent's model repository would
result in the rclone path not being cleaned-up. This leads to the PVC filling up
and requiring end-user manual intervention.

This change fixes this, by cleaning the rclone path even when errors occur.