fix(operator): Use StatefulSet.Spec.Replicas as reference for Server status
Before this change, we decide to mark the Server StatefulSet as ready by
comparing the `StatefulSet.Status.ReadyReplicas` with
`StatefulSet.Status.Replicas`. When the two are equal, we mark the StatefulSet
as "Ready".

However, this doesn't account for the possibility of `StatefulSet.Spec.Replicas`
to differ from `StatefulSet.Status.Replicas` (desired state vs actual state).

The change corrects this so that we only mark the StatefulSet as ready when
the number of ready replicas matches the desired state.