feat(k6): add scenario with multiple stages ramping up/down RPS
The added scenario allows one to configure an arbitrary number of stages,
with each consisting of a linear ramp-up/down to the desired rate and a
hold/plateau time.

Within each stage, the duration for which the inference RPS is held constant
is configured via one element in the `CONSTANT_RATE_DURATIONS_SECONDS`
environment variable (a vector of comma separated values), with the ramp-up/
down duration preceding it being 1/3rd of the hold time.