fix(operator): Add Status.AvailableReplicas to Model CRD (#5873)
* fix(operator): Add Status.AvailableReplicas to Model CRD

The Model.Status.Replicas field always reflects the total number of Model
replicas in existence, irrespective of their status. This is because it
is used by the scale subresource (enabling the configuration of HPA on
Model CRs).

However, we need a way of showing how many of those Replicas are available in
order to distinguish between various partial availability scenarios, when the
ModelReady condition is set to false.

This introduces a new Status.AvailableReplicas field, showing exactly that

* Add number of available replicas to Model printcolumns

* adjust log message to include min replicas

* update availableReplicas by using len of Assignment

---------

Co-authored-by: Sherif Akoush <sa@seldon.io>