refactor(Agent): More intuitive naming for the agent's main types
This is the first commit in a series, refactoring Agent code with a focus on code
legibility, cleanup and the use of modern golang idioms.

** It shouldn't introduce _any_ new functionality or change of behaviour. **

The hope is to slowly update names within the codebase and location of files so
as to make the code easier to understand and help with onboarding new developers
and contributors.

Main type renames:
- `Client`         -> `AgentServiceManager`
- `ClientSettings` -> `AgentServiceConfig`
- `ClientServices` -> `StorageManager`

The names of the files in which those types are defined have been kept the same
as to not confuse git's renaming detection (due to the large number of changes).
The files will be renamed in a dedicated, subsequent commit.

Main function renames:
- `Start` -> `StartControlLoop`
- `Stop`  -> `StopControlLoop`

Main field renames:
- `replicaConfig` -> `inferenceServerConfig`
- `settings`      -> `agentConfig`
- `conn`          -> `schedulerConn`

Other changes:

* Since Go 1.22, the semantics of for loop variables has changed:
https://tip.golang.org/doc/go1.22#language, with each iteration creating new
variables. Therefore, spurious within-loop asignments of the `iter := iter` type
are no longer useful.
    - make changes within the code to remove spurious assignments
    - add `copyloopvar` to the list of linters so as to detect when the old
    idiom is used
* Move from the `interface{}` type to the recommended `any` alias
* In tests, move from `NewGomegaWithT(t)` (deprecated) to `NewWithT(t)`
* Remove unused field `configChan` from `Client` (now `AgentServiceManager`)