fix(Agent): All subservices to start within maxElapsedTime*
Context
-------

The agent receives, via the cli or env var, two arguments controlling how much time
the agent should wait for its various subservices to start:

* `max-elapsed-time-ready-sub-service-before-start-minutes`, controlling the
amount of time the agent will wait for sub-services when it first starts (before
giving up and stopping). The default is 15 minutes, allowing for related
containers within the same pod such as `mlserver` or `triton` to become ready
(image pull, service startup)
* `max-elapsed-time-ready-sub-service-after-start-seconds`, controlling the time
the agent will wait for subservices to be ready on subsequent, periodic status
checks. The default is 30 seconds, and it is assumed that most of the time the
subservices will already be ready.

The list of subservices that the Agent waits for and periodically checks:
- Model Repository service (rclone)
- Inference Server (mlserver/triton)
- HTTP reverse proxy service (local to agent): provides HTTP endpoints targeted by
  the Load Balancer (envoy)
- gRPC reverse proxy service (local to agent): provides gRPC endpoints targeted
  by the Load Balancer (envoy)
- Agent Debug service (local to agent): provides an endpoint allowing the
  inspection of internal agent state
- Model Scaling service: allows the agent to signal model scale up/down needs
  based on per-model lag (outstanding requests)
- Drainer service: allows the agent reverse proxies to shutdown cleanly, without
  end-user perceived unavailability. On shutdown, the proxies will block new
  connections but still serve (drain) the outstanding ones before terminating

Issues
------

- Previous implementation waits on startup for a maximum of 15 min (by
default) **for each subservice** rather than as a total time. This means that
configuration issues in multiple external subservices (missing rclone secrets,
a custom inference server image that fails to pull or start) may create the
impression that the agent "hangs" while doing nothing.
- (related to above): The subservices are waited for in a sequential manner
(both on first start and subsequent checks)
- All subservice failures are considered fatal, irrespective of the importance
of the subservice to the end-user or to impact on availability (take the Agent
Debug service as an example)

Fixes
-----

- Parallel start-up and periodic health checks of all subservices
- The `max-elapsed-time-ready-*` variables are now considered total maximum wait
  times for all subservices. This is implemented by having a context that
  expires once the max time elapses. Any retries (to start the subservice, to
  check for its state) are also stopped once the context expires.
- Each subservice gets assigned a type from the following list (examples in brackets):
  * CriticalControlPlaneService (Drainer service)
  * CriticalDataPlaneService    (The HTTP/gRPC reverse proxy services)
  * AuxControlPlaneService      (Model Repository service)
  * AuxDataPlaneService         (Model Scaling service)
  * OptionalService             (Agent Debug service)
- According to newly implemented logic, the agent is considered in a failed
  state (and stopped) if on first startup any non-optional subservice fails.
  After start-up (startup is considered completed after the first successful
  connection to the scheduler), the agent is considered in a failed state if
  any critical subservice transitions in a failed state and is not able to
  recover during retries.

Tests
-----

A matrix-style test, `TestAgentReadiness` has been added. It checks for correct
behaviour under all possible combinations of failure/no failure for subservice types
("critical", "aux" and "optional") -- 8 cases in total.

We leverage parallel execution of test cases for `TestAgentReadiness` in order
to limit the additional test time in CI/CD pipelines as well as in running local
dev tests (<3 seconds added)