feat(Agent): Add readiness service, exposing endpoints for probing
The readiness service exposes an agent endpoint where the agent current status
can be verified. On calling the endpoint, the readiness service returns code 200
if the Agent container can be considered "ready", and an error code (404 or 503)
otherwise.

The key is that we start the readiness service very early in the Agent
lifecycle, even before an AgentServiceManager object exists. In this way,
readiness probes can start receiving answers immediately after the agent
container starts.

Currently, the Agent process stops if the control loop stops (typically, caused
by a failure in one of the critical subservices). We set up the code
so that in the future we can decide to restart subservices and the control loop,
while keeping the `/ready` endpoint responses in sync. This is done by resetting
the `ReadyService` target when `AgentServiceManager`'s ControlLoop stops.