feat(Agent): add readiness probes (#6560)
**What this PR does / why we need it**:

At the moment, using Core 2 in conjunction with k8s Pod Disruption Budgets (PDBs) doesn't work as expected, because on the inference server pod side, the pod is considered "Ready" from a PDB perspective while the agent is in draining mode.

This means that while we cleanly tear down an inference server (blocking new connections but continuing to serve all outstanding requests), PDB may continue to terminate other inference server replicas without affecting the disruption budget.

The consequence of this is that in the extreme, all inference server replicas serving particular models may be terminated (leading to unavailability), while also fulfilling all the PDB constraints.

As part of the changes, we refactor Agent code for readability and improve the way we wait and deal with internal and external dependent services.

**Which issue(s) this PR fixes**:
Fixes #INFRA-1392 (internal, add agent readiness probes)

** Details **

* refactor(Agent): More intuitive naming for the agent's main types

This is the first commit in a series, refactoring Agent code with a focus on code
legibility, cleanup and the use of modern golang idioms.

** It shouldn't introduce _any_ new functionality or change of behaviour. **

The hope is to slowly update names within the codebase and location of files so
as to make the code easier to understand and help with onboarding new developers
and contributors.

Main type renames:
- `Client`         -> `AgentServiceManager`
- `ClientSettings` -> `AgentServiceConfig`
- `ClientServices` -> `StorageManager`

The names of the files in which those types are defined have been kept the same
as to not confuse git's renaming detection (due to the large number of changes).
The files will be renamed in a dedicated, subsequent commit.

Main function renames:
- `Start` -> `StartControlLoop`
- `Stop`  -> `StopControlLoop`

Main field renames:
- `replicaConfig` -> `inferenceServerConfig`
- `settings`      -> `agentConfig`
- `conn`          -> `schedulerConn`

Other changes:

* Since Go 1.22, the semantics of for loop variables has changed:
https://tip.golang.org/doc/go1.22#language, with each iteration creating new
variables. Therefore, spurious within-loop asignments of the `iter := iter` type
are no longer useful.
    - make changes within the code to remove spurious assignments
    - add `copyloopvar` to the list of linters so as to detect when the old
    idiom is used
* Move from the `interface{}` type to the recommended `any` alias
* In tests, move from `NewGomegaWithT(t)` (deprecated) to `NewWithT(t)`
* Remove unused field `configChan` from `Client` (now `AgentServiceManager`)

* refactor(Agent): Rename the agent files to match type name changes

* fix(Agent): All subservices to start within maxElapsedTime*

Context
-------

The agent receives, via the cli or env var, two arguments controlling how much time
the agent should wait for its various subservices to start:

* `max-elapsed-time-ready-sub-service-before-start-minutes`, controlling the
amount of time the agent will wait for sub-services when it first starts (before
giving up and stopping). The default is 15 minutes, allowing for related
containers within the same pod such as `mlserver` or `triton` to become ready
(image pull, service startup)
* `max-elapsed-time-ready-sub-service-after-start-seconds`, controlling the time
the agent will wait for subservices to be ready on subsequent, periodic status
checks. The default is 30 seconds, and it is assumed that most of the time the
subservices will already be ready.

The list of subservices that the Agent waits for and periodically checks:
- Model Repository service (rclone)
- Inference Server (mlserver/triton)
- HTTP reverse proxy service (local to agent): provides HTTP endpoints targeted by
  the Load Balancer (envoy)
- gRPC reverse proxy service (local to agent): provides gRPC endpoints targeted
  by the Load Balancer (envoy)
- Agent Debug service (local to agent): provides an endpoint allowing the
  inspection of internal agent state
- Model Scaling service: allows the agent to signal model scale up/down needs
  based on per-model lag (outstanding requests)
- Drainer service: allows the agent reverse proxies to shutdown cleanly, without
  end-user perceived unavailability. On shutdown, the proxies will block new
  connections but still serve (drain) the outstanding ones before terminating

Issues
------

- Previous implementation waits on startup for a maximum of 15 min (by
default) **for each subservice** rather than as a total time. This means that
configuration issues in multiple external subservices (missing rclone secrets,
a custom inference server image that fails to pull or start) may create the
impression that the agent "hangs" while doing nothing.
- (related to above): The subservices are waited for in a sequential manner
(both on first start and subsequent checks)
- All subservice failures are considered fatal, irrespective of the importance
of the subservice to the end-user or to impact on availability (take the Agent
Debug service as an example)

Fixes
-----

- Parallel start-up and periodic health checks of all subservices
- The `max-elapsed-time-ready-*` variables are now considered total maximum wait
  times for all subservices. This is implemented by having a context that
  expires once the max time elapses. Any retries (to start the subservice, to
  check for its state) are also stopped once the context expires.
- Each subservice gets assigned a type from the following list (examples in brackets):
  * CriticalControlPlaneService (Drainer service)
  * CriticalDataPlaneService    (The HTTP/gRPC reverse proxy services)
  * AuxControlPlaneService      (Model Repository service)
  * AuxDataPlaneService         (Model Scaling service)
  * OptionalService             (Agent Debug service)
- According to newly implemented logic, the agent is considered in a failed
  state (and stopped) if on first startup any non-optional subservice fails.
  After start-up (startup is considered completed after the first successful
  connection to the scheduler), the agent is considered in a failed state if
  any critical subservice transitions in a failed state and is not able to
  recover during retries.

Tests
-----

A matrix-style test, `TestAgentReadiness` has been added. It checks for correct
behaviour under all possible combinations of failure/no failure for subservice types
("critical", "aux" and "optional") -- 8 cases in total.

We leverage parallel execution of test cases for `TestAgentReadiness` in order
to limit the additional test time in CI/CD pipelines as well as in running local
dev tests (<3 seconds added)

* feat(Agent): Add readiness service, exposing endpoints for probing

The readiness service exposes an agent endpoint where the agent current status
can be verified. On calling the endpoint, the readiness service returns code 200
if the Agent container can be considered "ready", and an error code (404 or 503)
otherwise.

The key is that we start the readiness service very early in the Agent
lifecycle, even before an AgentServiceManager object exists. In this way,
readiness probes can start receiving answers immediately after the agent
container starts.

Currently, the Agent process stops if the control loop stops (typically, caused
by a failure in one of the critical subservices). We set up the code
so that in the future we can decide to restart subservices and the control loop,
while keeping the `/ready` endpoint responses in sync. This is done by resetting
the `ReadyService` target when `AgentServiceManager`'s ControlLoop stops.

* feat(k8s): Add readiness and startup probes to inference server agents