fix(ci): kustomize version environment var
The github "draft-release" action defined an environment
variable (`KUSTOMIZE_VERSION`) that is now also used by various Core 2
makefiles. However, within the ci action steps expected the version to be
provided without a leading "v" (i.e not v5.2.1 but just 5.2.1) whereas the
makefiles expected the version with the leading v.

The implemented fix adjusts the name of the variable used directly by the
ci action.