fix(scheduler): Allow for pipelines with some of their statuses set to PipelineStatusUnknown (#6853)
* fix(scheduler): Allow for pipelines with some of their statuses set to PipelineStatusUnknown

When a new version of the scheduler (2.10) starts with the persistent store (on
PVC) from an old scheduler (<2.10), it loads the data about pipelines, and newly
added pipeline state fields are set to their default zero values.

In particular, `State.PipelineGwStatus` for all the existing pipelines in the
old cluster is set to `PipelineStatusUnknown`.

This was not considered as part of the "Running" statuses when deciding what
pipelines to consider as part of rebalancing towards pipeline-gw.