fix(gw-*): do not rely on settings in SeldonConfig.config.kafkaConfig.topics
Before Core 2.11, keep the number of partitions and replication factor as
microservice environment variables. Those settings are also present in the
newly introduced ScalingConfig, but they have no effect atm.

This is to ensure a smooth upgrade path, and make it possible in the future
to react to changes in those variables by watching the ConfigMap.