fix(probes): improve timing of k8s lifecycle probes (#6861)
Previous timings were leading to spurious restarts, especially on start-up as components waited to connect to the scheduler. It was possible for the scheduler image pull to take sufficiently long (and be timed in an unlucky way), such that other components restarted a number of times. The new timings simply reduce that possibility by being more generous wrt the startup probe.